fix overflows in reduction tests

+++ cuda/test/base/kernel_launch.cu
@@ -380,7 +380,7 @@ void run2d_row_reduction(std::shared_ptr<gko::CudaExecutor> exec)
             gko::Array<int64> output{exec, host_ref};
             for (int64 i = 0; i < num_rows; i++) {
                 host_ref.get_data()[2 * i] =
-                    num_cols * (num_cols + 1) * (i + 1);
+                    static_cast<int64>(num_cols) * (num_cols + 1) * (i + 1);
             }
 
             gko::kernels::cuda::run_kernel_row_reduction(
@@ -427,7 +427,8 @@ void run2d_col_reduction(std::shared_ptr<gko::CudaExecutor> exec)
                                        static_cast<size_type>(num_cols)};
             gko::Array<int64> output{exec, static_cast<size_type>(num_cols)};
             for (int64 i = 0; i < num_cols; i++) {
-                host_ref.get_data()[i] = num_rows * (num_rows + 1) * (i + 1);
+                host_ref.get_data()[i] =
+                    static_cast<int64>(num_rows) * (num_rows + 1) * (i + 1);
             }
 
             gko::kernels::cuda::run_kernel_col_reduction(
